transport: allow cred/cert callbacks to return >0
Allow credential and certificate checking callbacks to return > 0,
indicating that they did not want to act.  This was documented in the
credential acquisition callback already, but only loosely implemented.
Expand this to being supported in the http and ssh callbacks.
Additionally, enable the same mechanism for certificate validation.

This is useful to disambiguate any meaning for returning > 0 in the
publicly exposed credential and certificate functions
(`git_transport_smart_certificate_check`) and
(`git_transport_smart_credentials`).  But it may be more generally
useful for callers to be able to defer back to libgit2.